backend-api/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
baidusearch==1.0.3
beartype==0.12.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.24
botocore==1.37.24
browser-use==0.1.40
browsergym==0.13.3
browsergym-assistantbench==0.13.3
browsergym-core==0.13.3
browsergym-experiments==0.13.3
browsergym-miniwob==0.13.3
browsergym-visualwebarena==0.13.3
browsergym-webarena==0.13.3
browsergym-workarena==0.4.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.4.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
duckduckgo_search==7.5.5
english-words==2.0.1
evaluate==0.4.3
Faker==37.1.0
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
googlesearch-python==1.3.0
greenlet==3.1.1
gymnasium==1.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.49
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.21
lazy_loader==0.4
libvisualwebarena==0.0.15
libwebarena==0.0.4
loguru==0.7.3
lxml==5.3.1
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.5.0
monotonic==1.6
mpmath==1.3.0
multidict==6.3.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
ollama==0.4.7
openai==1.66.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.51.0
pluggy==1.5.0
portalocker==3.1.1
posthog==3.23.0
primp==0.14.0
propcache==0.3.1
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
text-generation==0.7.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
types-requests==2.32.0.20250328
types-tqdm==4.67.0.20250319
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
unidiff==0.7.5
urllib3==2.3.0
uvicorn==0.34.0
weblinx==0.3.2
weblinx-browsergym==0.0.1.dev14
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
